--- App.code-workspace
@@ -1,7 +1,0 @@
-{
-    "folders": [
-        {
-            "path": "."
-        }
-    ]
-}
