--- App.tsx
@@ -1,12 +1,0 @@
-
-import React from 'react';
-
-import { HomeSYS } from './src/modules/app';
-
-export default class App extends React.Component
-{
-    render()
-    {
-        return <HomeSYS></HomeSYS>;
-    }
-}
