--- app.json
@@ -1,37 +1,0 @@
-{
-    "expo": {
-        "name": "HomeSYS",
-        "slug": "HomeSYS",
-        "privacy": "unlisted",
-        "sdkVersion": "35.0.0",
-        "platforms": [
-            "android",
-            "web"
-        ],
-        "version": "1.0.0",
-        "orientation": "portrait",
-        "icon": "./assets/icon.png",
-        "splash": {
-            "image": "./assets/splash.png",
-            "resizeMode": "contain",
-            "backgroundColor": "#000000"
-        },
-        "updates": {
-            "fallbackToCacheTimeout": 0
-        },
-        "assetBundlePatterns": [
-            "**/*"
-        ],
-        "ios": {
-            "supportsTablet": true
-        },
-        "packagerOpts": {
-            "sourceExts": [
-                "ts",
-                "tsx",
-                "css",
-                "js"
-            ]
-        }
-    }
-}
