--- jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    preset: 'react-native',
-    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-};
