+++ src/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        <title>HomeSYS Dashboard</title>
+    </head>
+
+    <body>
+        <article id="page"></article>
+        <script src="./homesysapp.js"></script>
+    </body>
+
+</html>
